refactor: convert tests to ESM
